% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carsimr.geisler.R
\name{as_carsimrlist}
\alias{as_carsimrlist}
\title{as.carsimr}
\usage{
as_carsimrlist(x, ...)
}
\arguments{
\item{x}{A numeric matrix}

\item{...}{Optional.}
}
\description{
Converts matrices to class "carsimr"
}
