% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carsimr.geisler.R
\name{carsimr_convert}
\alias{carsimr_convert}
\title{c++ functions return a list of matrices, we want to make a carsimr object
that looks like the one we made originally.}
\usage{
carsimr_convert(x)
}
\arguments{
\item{x}{a list of matrices.}
}
\description{
c++ functions return a list of matrices, we want to make a carsimr object
that looks like the one we made originally.
}
