% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carsimr.geisler.R
\name{col.number}
\alias{col.number}
\title{col.number}
\usage{
col.number(x, y)
}
\arguments{
\item{x}{An element of a matrix.}

\item{y}{The matrix from which x is located.}
}
\description{
col.number
}
