% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_and_shuffle}
\alias{generate_and_shuffle}
\title{generate_and_shuffle}
\usage{
generate_and_shuffle(n, n2, p)
}
\arguments{
\item{n}{an integer with the number of things we need to simulate}

\item{n2}{an integer with the number of non-zero elements}

\item{p}{the probability of selecting a one (versus a two)}
}
\description{
Try to shuffle things
}
