% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carsimr.geisler.R
\name{initialize_grid}
\alias{initialize_grid}
\title{initialize_grid}
\usage{
initialize_grid(rho, r, c, p)
}
\arguments{
\item{rho}{The proportion of cars to fill the grid rounded to the nearest
whole number}

\item{r}{The number of rows in the grid}

\item{c}{The number of columns in the grid}

\item{p}{The proportion of blue and red cars as determined by the Bernoulli
distribution}
}
\description{
Initializes the grid of red and blue cars.
}
\details{
Using a specified proportion and grid size, we can randomly generate a matrix
of red and blue cars using the Bernoulli distribution. The seed has been
set for reproduceability.
}
\examples{
init_grid <- initialize_grid(rho = 0.3, r = 10, c = 10, p = 0.5)
}
