% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initialize_grid_cpp}
\alias{initialize_grid_cpp}
\title{C++ version of initialize_grid.}
\usage{
initialize_grid_cpp(rho, r, c, p, seed)
}
\arguments{
\item{rho}{The proportion of cars to fill the grid rounded to the nearest
whole number}

\item{r}{The number of rows in the grid}

\item{c}{The number of columns in the grid}

\item{p}{The proportion of blue and red cars as determined by the Bernoulli
distribution}
}
\description{
C++ version of initialize_grid.
}
