% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carsimr.geisler.R
\name{move_blue_cars}
\alias{move_blue_cars}
\title{move_blue_cars}
\usage{
move_blue_cars(m)
}
\arguments{
\item{m}{A matrix to be passed onto the function. The end result will show
the movement of blue cars (1's).}
}
\description{
move_blue_cars
}
