% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{move_blue_cpp}
\alias{move_blue_cpp}
\title{move_blue_cpp
C++ version of move_blue_cars.}
\usage{
move_blue_cpp(m)
}
\arguments{
\item{m}{A matrix of blue and red cars.}
}
\description{
move_blue_cpp
C++ version of move_blue_cars.
}
