% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carsimr.geisler.R
\name{move_cars}
\alias{move_cars}
\title{move_cars}
\usage{
move_cars(rho, r, c, p, trials)
}
\arguments{
\item{rho}{the proportion of cars to fill the grid, rounded to the nearest
whole number.}

\item{r}{the number of rows in the grid.}

\item{c}{the number of columns in the grid.}

\item{p}{the proportion of blue cars. 1-p represents the proportion of red
cars.}

\item{trials}{the number of simulations to be run on the grid.}
}
\description{
Provides a number of simulations of red & blue car movement on the grid.
}
\details{
On odd iterations, blue cars will be moved vertically. On even iterations,
red cars will be moved horizontally. Each iteration will be plotted as a
new matrix so that the car movement can be visualized.
}
\examples{
car_simulation <- move_cars(rho = 0.4, r = 9, c = 9, p = 0.7, trials = 10)
}
