% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carsimr.geisler.R
\name{move_cars_cpp}
\alias{move_cars_cpp}
\title{Wrapper to c++ functions.}
\usage{
move_cars_cpp(rho, r, c, p, trials)
}
\arguments{
\item{rho}{A proportion between 0 and 1 of the number of grids that
should be filled with cars. Alternatively, an integer value specifying
the precise number of cars.}

\item{r}{The number of rows in the car grid.}

\item{c}{The number of columns in the car grid.}

\item{p}{The probability (between 0 and 1) of selecting a blue car.
This means the probability of selecting a red car is 1-p.}

\item{trials}{The number of grid moves to store.}
}
\description{
Wrapper to c++ functions.
}
