% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{move_cars_cpp_c}
\alias{move_cars_cpp_c}
\title{move_cars_cpp_c}
\usage{
move_cars_cpp_c(rho, r, c, p, trials)
}
\arguments{
\item{rho}{The proportion or number of cars in a grid.}

\item{r}{The number of rows in the grid.}

\item{c}{The number of columns in the grid.}

\item{p}{The proportion of red cars in the grid.}

\item{trials}{The number of grid movements to be simulated.}
}
\description{
C++ version of move_cars
}
