% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carsimr.geisler.R
\name{move_cars_par}
\alias{move_cars_par}
\title{move_cars_par}
\usage{
move_cars_par(rho, r, c, p, trials, replicates, cores)
}
\arguments{
\item{rho}{The proportion or number of cars on the grid.}

\item{r}{The number of rows in the grid.}

\item{c}{The number of columns in the grid.}

\item{p}{The proportion of red cars in the grid.}

\item{trials}{The number of grid simulations.}

\item{replicates}{The number of repeat grids to be created.}

\item{cores}{The number of cores to be devoted to the computations.}
}
\description{
move_cars_par
}
