% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carsimr.geisler.R
\name{move_red_cars}
\alias{move_red_cars}
\title{move_red_cars}
\usage{
move_red_cars(m)
}
\arguments{
\item{m}{A matrix to be passed onto the function. The end result will show
the movement of red cars (2's).}
}
\description{
move_red_cars
}
