% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carsimr.geisler.R
\name{plot.carsimr}
\alias{plot.carsimr}
\title{plot.carsimr}
\usage{
\method{plot}{carsimr}(x, y, ...)
}
\arguments{
\item{x}{A matrix of class "carsimr"}

\item{y}{Optional if x is complete}

\item{...}{A symbol meaning that other arguments can be used.}
}
\description{
This takes the matrix generated of class carsimr and converts it to an
image of red and blue cars on the grid.
}
