% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carsimr.geisler.R
\name{plot.carsimrlist}
\alias{plot.carsimrlist}
\title{plot.carsimrlist}
\usage{
\method{plot}{carsimrlist}(x, y, ...)
}
\arguments{
\item{x}{A list containing the matrices from each iteration.}

\item{y}{Ignored}

\item{...}{Ignored}
}
\description{
Plots each iteration of car movement. In total, there will be 'trials+1'
matrices returned.
}
