% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{randWrapper}
\alias{randWrapper}
\title{"inline" allows us to make copies of this function where we need it. This
is useful because random_shuffle() needs a dynamic function, not a static
vector, in order create a random number stream based on one from R.}
\usage{
randWrapper(n)
}
\description{
"inline" allows us to make copies of this function where we need it. This
is useful because random_shuffle() needs a dynamic function, not a static
vector, in order create a random number stream based on one from R.
}
