% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rho_value}
\alias{rho_value}
\title{rho_value}
\usage{
rho_value(x, row, col)
}
\arguments{
\item{x}{The proportion or number of cars to fill the matrix.}

\item{row}{The number of rows.}

\item{col}{The number of columns.}
}
\description{
Returns rounded value of rho that can be passed to the number_cars variable.
}
