% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carsimr.geisler.R
\name{row.number}
\alias{row.number}
\title{row.number}
\usage{
row.number(x, y)
}
\arguments{
\item{x}{An element of a matrix.}

\item{y}{The matrix from which x is located.}
}
\description{
row.number
}
