% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carsimr.geisler.R
\name{simulate_grid}
\alias{simulate_grid}
\title{simulate_grid}
\usage{
simulate_grid(matrix, move_red_cars, move_blue_cars, trials)
}
\arguments{
\item{matrix}{A matrix to be passed onto the function that will demonstrate
car movement.}

\item{move_red_cars}{Calls the 'move_red_cars' function notated above.}

\item{move_blue_cars}{Calls the 'move_blue_cars' function notated above.}

\item{trials}{Specifies the number of times the cars move on the grid.}
}
\description{
simulate_grid
}
