% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carsimr.geisler.R
\name{simulate_grid_cpp}
\alias{simulate_grid_cpp}
\title{simulate_grid_cpp}
\usage{
simulate_grid_cpp(m, trials)
}
\arguments{
\item{m}{a numeric matrix.}

\item{trials}{The number of car movement iterations.}
}
\description{
A wrapper function that converts simulate_grid_cpp to class "carsimr"
}
