% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_grid_cpp_c}
\alias{simulate_grid_cpp_c}
\title{simulate_grid_cpp}
\usage{
simulate_grid_cpp_c(m, trials)
}
\arguments{
\item{m}{A numeric matrix of class "carsimr" to be simulated.}

\item{trials}{Specifies the number of times the cars move on the grid.}
}
\description{
C++ version of simulate_grid.
}
