% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carsimr.geisler.R
\name{track_blocks}
\alias{track_blocks}
\title{track_blocks}
\usage{
track_blocks(moves)
}
\arguments{
\item{moves}{A list of class "carsimr.list".}
}
\description{
Tracks the proportion of blocked cars at each iteration. Returns a vector.
}
