% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_grid_cpp}
\alias{simulate_grid_cpp}
\title{simulate_grid_cpp}
\usage{
simulate_grid_cpp(m, trials)
}
\arguments{
\item{trials}{Specifies the number of times the cars move on the grid.}

\item{move_red_cars}{Calls the 'move_red_cars' function notated above.}

\item{move_blue_cars}{Calls the 'move_blue_cars' function notated above.}
}
\description{
C++ version of simulate_grid.
}
